requirements.txt
asgiref==3.8.1
Django==5.1.4
django-modeltranslation==0.19.12
pillow==11.1.0
sqlparse==0.5.3
